Factor out common stuff to hook
